Change Branch protection rules to make use of new Action
Disable "Dismiss stale pull request approvals when new commits are
pushed" because this will be done by the CI if anything changed.
Enable "Require status checks to pass before merging", to ensure that
the CI will complete before merging is enabled or reviews dismissed.

Also, I've granted myself ability to force push on main to not spam with
PR when testing CI and rebasing.

Signed-off-by: Bojan Potočnik <info@bojanpotocnik.com>