Ignore change hunk range information
If code is added or removed (much) after changes from the commit which
is being rebased, range information for specific chunk will stay the
same and therefore ignored.
If code is added or removed (much) before changes from the commit which
is being rebased, range information for the following chunks will change
and PR will be dismissed.

With this commit, the second case is made equal to the first case and
therefore PR is not dismissed in all instead of "half" cases.
Section heading is kept to prevent possibility of adding/removing the
code with rebase combined with moving the code from commit in a way to
keep the range information exactly the same.

Signed-off-by: Bojan Potočnik <info@bojanpotocnik.com>